backend/requirements.txt
Django~=3.0.11
djangorestframework~=3.12.2
django-filter~=2.4.0
django-rest-auth~=0.9.5
celery~=5.0.2
django-allauth~=0.44.0
PyMySQL~=0.10.1